<Update>:<Ability System 변경. Editor 적용완료. 화살 아이템과 유물 아이템 엔터티 추가작업>
<description>
 1. 화살 프리팹 추가 구현.
 2. 어빌리티 시스템 개선 진행중.
 3. 변경된 어빌리티 시스템 에디터 작업완료.
 4. 스프라이트 폴더 정리 완료.
 5. 추가적인 종류의 어빌리티 타입 추가중.

<issue-number>
# 연관된 이슈 첨부, 여러 개 추가 가능

+++ ArrowDefence_Project/Assets/02.Scripts/CodingCat_Script/01.Item Script/Item_Bow.cs
@@ -42,7 +42,7 @@
                     case ABILITY_TYPE.DAMAGE:
                         var damage = abilities[i] as AbilityDamage;
                         if (damage != null) {
-                            tempList.Add(new AbilityDamage(System.Convert.ToInt16(damage.GetCount())));
+                            tempList.Add(new AbilityDamage(System.Convert.ToInt32(damage.GetCount())));
                         }
                         else {
                             CatLog.ELog("Ability Type Not Matched.");
@@ -58,7 +58,7 @@
                     case ABILITY_TYPE.CRITICALCHANCE:
                         var criticalChance = abilities[i] as AbilityCritChance;
                         if (criticalChance != null) {
-                            tempList.Add(new AbilityCritChance(System.Convert.ToByte(criticalChance.GetCount())));
+                            tempList.Add(new AbilityCritChance(System.Convert.ToInt32(criticalChance.GetCount())));
                         }
                         else {
                             CatLog.ELog("Ability Type Not Matched.");
